comecando o pedido controller

+++ loja/src/main/java/com/mestredagrelha/loja/dto/ProdutoRequestDTO.java
@@ -2,6 +2,6 @@ package com.mestredagrelha.loja.dto;
 
 public record ProdutoRequestDTO(
         Integer produto_id, Integer categoria_id, String nome,
-        String descricao, double preco, Integer estoque, String imagem
+        String descricao, Integer preco, Integer estoque, String imagem
 ) {
 }
